==> 1_hw_MLiB.ipynb <==
Автор материала: Зраев Артем.

Можно использовать в каких угодно целях.

<b> В задании нужно загрузить датасет с данными оттока и ответить на несколько вопросов (написать код). При этом сам датасет уже есть и его необязательно качать с репозитория</b>

Цель задания: проверить базовые навыки работы студентов с Pandas, умение проводить такой же базовый EDA (exploratory data analysis), делать feature engineering и обучать и валидировать модель.

Список столбцов с типами данных в датасете:

- customerID           object
- gender               object
- SeniorCitizen         int64
- Partner              object
- Dependents           object
- tenure                int64
- PhoneService         object
- MultipleLines        object
- InternetService      object
- OnlineSecurity       object
- OnlineBackup         object
- DeviceProtection     object
- TechSupport          object
- StreamingTV          object
- StreamingMovies      object
- Contract             object
- PaperlessBilling     object
- PaymentMethod        object
- MonthlyCharges      float64
- TotalCharges         object
- Churn                object

import pandas as pd
import numpy as np

df = pd.read_csv("./WA_Fn-UseC_-Telco-Customer-Churn.csv")
df.head(3)

##### 1. Какое соотношение мужчин и женщин в представленном наборе данных?

df['gender'].value_counts(normalize = True)*100

##### 2. Какое количество уникальных значений у поля InternetService?

df['InternetService'].nunique()

##### 3. Выведите статистики по полю TotalCharges (median, mean, std).

df['TotalCharges'].describe()

В чем странность того, что вы получили? (подсказка: смотреть нужно на тип данных)

Т.к. тип данных - object, метод describe выводит результаты, отличающиеся от результатов для количественных типов данных

Но в данном случае, тип данных можно скорректировать, предварительно обработав пустые строки

#Можно сразу проебразовать пустые строки и перевести столбец в тип float, но выполним это задание далее

#df.loc[df['TotalCharges'] == ' ', 'TotalCharges'] = 0

#df['TotalCharges'] = df['TotalCharges'].astype('float')

#df['TotalCharges'].describe()

##### 4. Сделайте замену значений поля PhoneService  на числовые (Yes->1, No->0)

# Выведем все значения признака PhoneService
df['PhoneService'].unique()

df['PhoneService'].replace({'Yes':1, 'No': 0}, inplace=True)

# Проверим получившееся преобразование
df['PhoneService'].unique()

##### 5. Сделайте замену пробелов в поле TotalCharges на np.nan и приведите поле к типу данных float32. Затем заполните оставшиеся пропуски значением 0 с помощью метода fillna у столбца. Снова выведите статистики и сравните с тем, что вы видели в вопросе 3

df.loc[df['TotalCharges'] == ' ', 'TotalCharges'] = np.nan

df['TotalCharges'].isna().sum()

df['TotalCharges'].fillna(0, inplace=True)

# Перепроверим, остались ли пропуски
df['TotalCharges'].isna().sum()

##### 6. Сделайте замену значений поля Churn на числовые (Yes -> 1, No - 0)

# Выведем все уникальные значения
df['Churn'].unique()

df['Churn'].replace({'Yes': 1, 'No': 0}, inplace=True)

df['Churn'].unique()

##### 7. Сделайте замену значений полей StreamingMovies, StreamingTV, TechSupport  на числовые (Yes -> 1, No -> 0, No internet service->0)

#Выведем все уникальные значения указанных в задании признаков (вдруг есть еще какие-то значения, которые не указаны)

features = ['StreamingMovies', 'StreamingTV', 'TechSupport']
for feature in features:
    print(f'Уникальные значения {feature} : {df[feature].unique()}')

for feature in features:
    df[feature].replace({'Yes': 1, 'No': 0, 'No internet service': 0}, inplace=True)

# Проверим корректность преобразования 
for feature in features:
    print(f'Уникальные значения {feature} : {df[feature].unique()}')

##### 8. Заполните пропуски в поле PhoneService значением 0

df['PhoneService'].isnull().sum()

В признаке PhoneService пропусков не найдено

#### 8. Для нашего датасета оставьте только указанный ниже список полей, удалив все другие и выведите верхние 3 строки

df.columns

columns = ['gender', 'tenure', 'PhoneService', 'TotalCharges', 
           'StreamingMovies', 'StreamingTV', 'TechSupport', 'Churn']
#Ваш код здесь



# Создадим новый датасет, содержащий указанные признаки df
df_new = df[columns]
df_new.head(3)

##### 9. Разделите датасет на тренировочную и тестовую выборку (подсказка - воспользуйтесь train_test_split из sklearn.model_selection. Ссылка - https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.train_test_split.html)

from sklearn.model_selection import train_test_split

features = ['gender', 'tenure', 'PhoneService', 'TotalCharges', 'StreamingMovies', 'StreamingTV', 'TechSupport']
target = 'Churn'
#Ваш код здесь


X_train, X_test, y_train, y_test = train_test_split(df_new[features], df_new[target], test_size = 0.25, random_state = 1)


##### 10. соберите pipeline для поля gender (нужно разобраться и изучить https://scikit-learn.org/stable/modules/generated/sklearn.pipeline.Pipeline.html ) из классов ColumnSelector и OHEEncoder, которые уже написаны ниже заранее

from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.pipeline import Pipeline

class ColumnSelector(BaseEstimator, TransformerMixin):
    """
    Transformer to select a single column from the data frame to perform additional transformations on
    """
    def __init__(self, key):
        self.key = key

    def fit(self, X, y=None):
        return self

    def transform(self, X):
        return X[self.key]
    
class NumberSelector(BaseEstimator, TransformerMixin):
    """
    Transformer to select a single column from the data frame to perform additional transformations on
    Use on numeric columns in the data
    """
    def __init__(self, key):
        self.key = key

    def fit(self, X, y=None):
        return self

    def transform(self, X):
        return X[[self.key]]
    
class OHEEncoder(BaseEstimator, TransformerMixin):
    def __init__(self, key):
        self.key = key
        self.columns = []

    def fit(self, X, y=None):
        self.columns = [col for col in pd.get_dummies(X, prefix=self.key).columns]
        return self

    def transform(self, X):
        X = pd.get_dummies(X, prefix=self.key)
        test_columns = [col for col in X.columns]
        for col_ in test_columns:
            if col_ not in self.columns:
                X[col_] = 0
        return X[self.columns]

gender = Pipeline([
                ('selector', ColumnSelector(key='gender')),
                ('ohe', OHEEncoder(key='gender'))
            ])

from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.pipeline import Pipeline


# 'key' заменила на 'column'
class ColumnSelector(BaseEstimator, TransformerMixin):
    """
    Transformer to select a single column from the data frame to perform additional transformations on
    Класс для выбора отдного признака из датафрейма для дальнейших преобразований
    """
    def __init__(self, column):
        self.column = column

    def fit(self, X, y=None):
        return self

    def transform(self, X):
        return X[self.column]
    
class NumberSelector(BaseEstimator, TransformerMixin):
    """
    Transformer to select a single column from the data frame to perform additional transformations on
    Use on numeric columns in the data
    
    Используется для числовых признаков
    """
    def __init__(self, column):
        self.column = column

    def fit(self, X, y=None):
        return self

    def transform(self, X):
        return X[[self.column]]
    
class OHEEncoder(BaseEstimator, TransformerMixin):
    
    """
    Делаем преобразование категориальных признаков (в данном случае в дамми)
    """
    
    def __init__(self, column):
        self.column = column
        self.columns = []
    
    def fit(self, X, y=None):
        self.columns = [col for col in pd.get_dummies(X, prefix=self.column).columns]
        return self
    
    #Формируется список из новых колонок (дамми) и возвращается датасет с новыми признаками
    def transform(self, X):
        X = pd.get_dummies(X, prefix=self.column)
        test_columns = [col for col in X.columns]
        for col_ in test_columns:
            if col_ not in self.columns:
                X[col_] = 0
        return X[self.columns]

gender = Pipeline([
                ('selector', ColumnSelector(column='gender')),
                ('ohe', OHEEncoder(column='gender'))
            ])

##### 11. Вызовите метод fit_transform у пайплайна gender и передайте туда нашу тренировочную выборку (пример по ссылке из документации https://scikit-learn.org/stable/modules/generated/sklearn.pipeline.Pipeline.html#sklearn.pipeline.Pipeline.fit)

gender.fit_transform(X_train)


##### 12. Здесь код писать уже не нужно (все сделано за вас). К полю tenure применяем StandardScaler (нормируем и центрируем). Ссылка - https://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.StandardScaler.html
Вопрос - в каких случаях это может быть полезно?

from sklearn.preprocessing import StandardScaler

tenure =  Pipeline([
                ('selector', NumberSelector(column='tenure')),
                ('standard', StandardScaler())
            ])

##### 13. Напишите аналогичный (как для tenure) преобразователь поля TotalCharges

TotalCharges = Pipeline([
                    ('selector', NumberSelector(column = 'TotalCharges')),
                    ('standard', StandardScaler())
            ])


Объединение всех "кубиков" очень легко сделать таким образом

from sklearn.pipeline import FeatureUnion

number_features = Pipeline([
                ('selector', ColumnSelector(column=['PhoneService',
                                                 'StreamingMovies', 'StreamingTV', 
                                                 'TechSupport']))
            ])

feats = FeatureUnion([('tenure', tenure),
                      ('TotalCharges', TotalCharges),
                      ('continuos_features', number_features),
                      ('gender', gender)])
feature_processing = Pipeline([('feats', feats)])

На этом этапе что мы сделали:
1. написали преобразователь поля gender, который делает OHE кодирование
2. написали преобразователь для поля tenure, который нормирует и центрирует его 
3. повторили п. 2 для поля TotalCharges
3. для всех остальных просто взяли признаки как они есть, без изменений

У нас уже готов наш пайплайн, который преобразовывает признаки. Давайте обучим модель поверх него. В качестве модели возьмем RandomForestClassifier

from sklearn.ensemble import RandomForestClassifier

pipeline = Pipeline([
    ('features',feats),
    ('classifier', RandomForestClassifier(random_state = 42)),
])

pipeline.fit(X_train, y_train)

##### 14. Сделайте прогноз вероятности оттока для X_test с помощью нашего предобученного на предыдущем шаге пайплайна и убедитесь что вам возвращаются вероятности для 2 классов

preds = pipeline.predict_proba(X_test)
preds.shape


X_test.shape[0] == preds.shape[0]

##### 15. Посчитайте метрики качества получившейся модели (roc_auc, logloss)

from sklearn.metrics import roc_auc_score, log_loss

#Ваш код здесь

roc_square = roc_auc_score(y_test, preds[:,1])
roc_square

# чем ближе к 1, тем лучше


logloss_metric = log_loss(y_test, preds[:,1])
logloss_metric

### Сохраним наш пайплайн

pip install dill

import dill


with open("model_RF.dill", "wb") as f:
    dill.dump(pipeline, f)
