==> Step2_predict.ipynb <==
### Проверка работоспособности и качества пайплайна

Здесь мы еще не запускаем никакое API, а загружаем модель (pipeline) напрямую и проверяем на отложенной (тестовой) выборке

import pandas as pd
from sklearn.metrics import roc_auc_score,roc_curve,scorer
import dill
dill._dill._reverse_typemap['ClassType'] = type

X_test = pd.read_csv("X_test.csv")
y_test = pd.read_csv("y_test.csv")

X_test.head(3)

with open('logreg_pipeline.dill', 'rb') as in_strm:
    pipeline = dill.load(in_strm)

predictions = pipeline.predict_proba(X_test)
pd.DataFrame({'preds': predictions[:, 1]}).to_csv("test_predictions.csv", index=None)

roc_auc_score(y_score=predictions[:, 1][:], y_true=y_test.iloc[:])
