==> 3_hw_MLiB.ipynb <==
### Домашнее задание

1. обучить несколько разных моделей на наборе данных ССЗ (train_case2.csv): логрег, бустинг, лес и т.д - на ваш выбор 2-3 варианта
2. при обучении моделей обязательно использовать кроссвалидацию
3. вывести сравнение полученных моделей по основным метрикам классификации: pr/rec/auc/f_score (можно в виде таблицы, где строки - модели, а столбцы - метрики)
4. сделать выводы о том, какая модель справилась с задачей лучше других
5. (опциональный вопрос) какая метрика (precision_recall_curve или roc_auc_curve) больше подходит в случае сильного дисбаланса классов? (когда объектов одного из классов намного больше чем другого). 

p.s.В вопросе проще разобраться, если вспомнить оси на графике roc auc curve и рассмотреть такой пример:

Имеется 100000 объектов, из которых только 100 - класс "1" (99900 - класс "0", соответственно). 
Допустим, у нас две модели:

- первая помечает 100 объектов как класс 1, но TP = 90
- вторая помечает 1000 объектов как класс 1, но TP такой же - 90

Какая модель лучше и почему? И что позволяет легче сделать вывод - roc_auc_curve или precision_recall_curve?

import numpy as np
import pandas as pd

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import cross_val_score, train_test_split
from sklearn.metrics import precision_recall_curve, roc_auc_score
from scipy.sparse import hstack
from sklearn.pipeline import Pipeline, make_pipeline
from sklearn.base import BaseEstimator, TransformerMixin

df = pd.read_csv('/Users/ekaterina/Desktop/LEARN/IT/Машинное обучение в бизнесе/lection3_materials/train_case2.csv', ';')
df.head(3)

#разделим данные на train/test
X_train, X_test, y_train, y_test = train_test_split(df.drop('cardio', 1), 
                                                    df['cardio'], random_state=0)

К полям:

gender, cholesterol применим OHE-кодирование

age, height, weight, ap_hi, ap_lo - standardScaler

gluc, smoke, alco, active - оставим пока как есть

class ColumnSelector(BaseEstimator, TransformerMixin):
    """
    Transformer to select a single column from the data frame to perform additional transformations on
    """
    def __init__(self, key):
        self.key = key

    def fit(self, X, y=None):
        return self

    def transform(self, X):
        return X[self.key]
    
class NumberSelector(BaseEstimator, TransformerMixin):
    """
    Transformer to select a single column from the data frame to perform additional transformations on
    Use on numeric columns in the data
    """
    def __init__(self, key):
        self.key = key

    def fit(self, X, y=None):
        return self

    def transform(self, X):
        return X[[self.key]]
    
class OHEEncoder(BaseEstimator, TransformerMixin):
    def __init__(self, key):
        self.key = key
        self.columns = []

    def fit(self, X, y=None):
        self.columns = [col for col in pd.get_dummies(X, prefix=self.key).columns]
        return self

    def transform(self, X):
        X = pd.get_dummies(X, prefix=self.key)
        test_columns = [col for col in X.columns]
        for col_ in test_columns:
            if col_ not in self.columns:
                X[col_] = 0
        return X[self.columns]


from sklearn.preprocessing import StandardScaler


continuos_cols = ['age', 'height', 'weight', 'ap_hi', 'ap_lo']
cat_cols = ['gender', 'cholesterol']
base_cols = ['gluc', 'smoke', 'alco', 'active']

continuos_transformers = []
cat_transformers = []
base_transformers = []

for cont_col in continuos_cols:
    transfomer =  Pipeline([
                ('selector', NumberSelector(key=cont_col)),
                ('standard', StandardScaler())
            ])
    continuos_transformers.append((cont_col, transfomer))
    
for cat_col in cat_cols:
    cat_transformer = Pipeline([
                ('selector', ColumnSelector(key=cat_col)),
                ('ohe', OHEEncoder(key=cat_col))
            ])
    cat_transformers.append((cat_col, cat_transformer))
    
for base_col in base_cols:
    base_transformer = Pipeline([
                ('selector', NumberSelector(key=base_col))
            ])
    base_transformers.append((base_col, base_transformer))

from sklearn.pipeline import FeatureUnion

feats = FeatureUnion(continuos_transformers+cat_transformers+base_transformers)
feature_processing = Pipeline([('feats', feats)])

feature_processing.fit_transform(X_train)

classifier = Pipeline([
    ('features',feats),
    ('classifier', LogisticRegression(random_state = 42)),
])


#запустим кросс-валидацию
cv_scores = cross_val_score(classifier, X_train, y_train, cv=16, scoring='roc_auc')
cv_score = np.mean(cv_scores)
cv_score_std = np.std(cv_scores)
print('CV score is {}+-{}'.format(cv_score, cv_score_std))

#обучим пайплайн на всем тренировочном датасете
classifier.fit(X_train, y_train)
y_score = classifier.predict_proba(X_test)[:, 1]

#### Задание 1. 
Обучить несколько разных моделей на наборе данных ССЗ (train_case2.csv): логрег, бустинг, лес и т.д - на ваш выбор 2-3 варианта 
#### Задание 2. 
При обучении моделей обязательно использовать кросс валидацию

from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier

rf_classifier = Pipeline([
    ('features',feats),
    ('rf_classifier', RandomForestClassifier(random_state = 42)),
])

#запустим кросс-валидацию
rf_cv_scores = cross_val_score(rf_classifier, X_train, y_train, cv=16, scoring='roc_auc')
rf_cv_score = np.mean(rf_cv_scores)
rf_cv_score_std = np.std(rf_cv_scores)
print('CV score is {}+-{}'.format(rf_cv_score, rf_cv_score_std))


gb_classifier = Pipeline([
    ('features',feats),
    ('gb_classifier', GradientBoostingClassifier(random_state = 42)),
])

#запустим кросс-валидацию
gb_cv_scores = cross_val_score(gb_classifier, X_train, y_train, cv=16, scoring='roc_auc')
gb_cv_score = np.mean(gb_cv_scores)
gb_cv_score_std = np.std(gb_cv_scores)
print('CV score is {}+-{}'.format(gb_cv_score, gb_cv_score_std))


Подберем гиперпараметры

n_estimators = [50, 70]
max_depth = [3, 5, 7]

for n in n_estimators:
    for depth in max_depth:
        gb_classifier = Pipeline([
            ('features',feats),
            ('gb_classifier', GradientBoostingClassifier(n_estimators = n, max_depth = depth, random_state = 42)),
        ])

        #запустим кросс-валидацию
        gb_cv_scores = cross_val_score(gb_classifier, X_train, y_train, cv=16, scoring='roc_auc')
        gb_cv_score = np.mean(gb_cv_scores)
        gb_cv_score_std = np.std(gb_cv_scores)
        print('n_estimators: {}, max_depth = {}, CV score is {}+-{}'.format(n, depth, gb_cv_score, gb_cv_score_std))

Лучшие параметры: n_estimators = 70, max_depth = 5 не существенно улучшили результат на кросс валидации. Добавим их в модель. Введем также дополнительный параметр min_samples_leaf

rf_classifier = Pipeline([
    ('features',feats),
    ('rf_classifier', RandomForestClassifier(n_estimators = 70, max_depth = 5, min_samples_leaf=15, random_state = 42)),
])

#запустим кросс-валидацию
rf_cv_scores = cross_val_score(rf_classifier, X_train, y_train, cv=16, scoring='roc_auc')
rf_cv_score = np.mean(rf_cv_scores)
rf_cv_score_std = np.std(rf_cv_scores)
print('CV score is {}+-{}'.format(rf_cv_score, rf_cv_score_std))


gb_classifier = Pipeline([
    ('features',feats),
    ('gb_classifier', GradientBoostingClassifier(n_estimators = 70, max_depth = 5, min_samples_leaf=15, random_state = 42)),
])

#запустим кросс-валидацию
gb_cv_scores = cross_val_score(gb_classifier, X_train, y_train, cv=16, scoring='roc_auc')
gb_cv_score = np.mean(gb_cv_scores)
gb_cv_score_std = np.std(gb_cv_scores)
print('CV score is {}+-{}'.format(gb_cv_score, gb_cv_score_std))

# Создадим функцию, обучающую pipeline и производящая оценку модели

def evaluate_prediction(classifier, X_train, y_train, X_test, y_test, b=1):
    classifier.fit(X_train, y_train)
    preds = classifier.predict_proba(X_test)[:, 1]
    
    #Определим метрики
    precision, recall, threshold = precision_recall_curve(y_test, preds)
    fscore = (1+b**2)*(precision * recall) / (b**2*precision + recall)
    ix = np.argmax(fscore)
    
    return precision[ix], recall[ix], fscore[ix], roc_auc_score(y_test, preds)

logreg_prec, logreg_rec, logreg_fscore, logreg_roc_auc = evaluate_prediction(classifier, X_train, y_train, X_test, y_test)
rf_prec, rf_rec, rf_fscore, rf_roc_auc = evaluate_prediction(rf_classifier, X_train, y_train, X_test, y_test)
gb_prec, gb_rec, gb_fscore, gb_roc_auc = evaluate_prediction(gb_classifier, X_train, y_train, X_test, y_test)

#### Задание 3. 

Вывести сравнение полученных моделей по основным метрикам классификации: pr/rec/auc/f_score (можно в виде таблицы, где строки - модели, а столбцы - метрики)

# Выведем таблицу с результатами

comparing_results = pd.DataFrame({
    'classifier': ['Logistic Regression', 'Random Forest', 'Gradient Boosting'],
    'precision': [logreg_prec, rf_prec, gb_prec],
    'recall': [logreg_rec, rf_rec, gb_rec],
    'f score': [logreg_fscore, rf_fscore, gb_fscore],
    'roc auc score': [logreg_roc_auc, rf_roc_auc, gb_roc_auc]
})

comparing_results

#### Задание 4. 
Сделать выводы о том, какая модель справилась с задачей лучше других

В данной задаче для нас метрика precision важнее, нежели recall (то есть в данной конкретной задаче стремимся минимизировать ложноотрицательные результаты - FN)

Самые лучше результаты в совокупсности по всем метрикам дала модель GradientBoosting. Чаще всего на практике эта модель действительно оказывается сильнее остальных. Подобранные гиперпараметры также позволили улучшить результат


#### Задание 5. 
Какой график (precision_recall_curve или roc_auc_curve) больше подходит в случае сильного дисбаланса классов? (когда объектов одного из классов намного больше чем другого, например, 1 к 1000).

Оси PR кривой - Precision и Recall

Оси ROC AUC кривой - TPR , FPR

В случае ROC AUC кривой - допустим, что количество 0 класса существенно превышает 1 класс. Тогда FPR будет очень долго оставаться на низком значении, в то время как TPR будет очень быстро расти. Это может привести к ложно оптимическому результату

В задачах с дисбалансом классов лучше использовать PR кривую
